% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_model.R
\name{fit_maxspeed}
\alias{fit_maxspeed}
\title{Fit time-dependent maximum speed model}
\usage{
fit_maxspeed(dat)
}
\arguments{
\item{dat}{\verb{[data.frame]} Subset of the maximum speed database. Requires columns \code{max_spd_kmh} and \code{mass_kg}.}
}
\value{
\verb{[nls]} A fitted non-linear regression model.
}
\description{
Fit time-dependent maximum speed model
}
