% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_model.R
\name{predict_maxspeed}
\alias{predict_maxspeed}
\title{Maximum speed predictions}
\usage{
predict_maxspeed(mass_kg, mod)
}
\arguments{
\item{mass_kg}{\verb{[numeric]} Mass of animals for model prediction (in kg)}

\item{mod}{\verb{[nls]} A fitted time-dependent maximum speed model (see \code{fit_maxspeed()})}
}
\value{
\verb{[numeric]} Predicted maximum speeds for animals of size \code{mass_kg}
}
\description{
Maximum speed predictions
}
